another refactor due to PR comments
