+++ aidial_adapter_openai/completions.py
